django/requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0,<4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests>=2.26.0,<3.0.0
cryptography==42.0.7
PyJWT==2.8.0
gunicorn==22.0.0
